Bug: Firefox unable to draw full image of certain mimetypes
introduced temporary fix that does not attempt to crop the video frames to match cropped video. Instead, frames are drawn straight from the source using the simplest overload for the drawImage method. This only seems to be a problem on firefox when trying to draw frames from unsupported mimetypes like ".mov"